+++ apps/web/src/features/safenet/components/SafenetTxStatusSteps/styles.module.css
@@ -7,3 +7,11 @@
   height: 16px;
   width: 16px;
 }
+
+.safenetLogo {
+  border-radius: 100px;
+  background-color: var(--color-background-main);
+  width: fit-content;
+  padding: 4px 8px;
+  margin-left: 32px;
+}
